Move `TestCommands` to `given` package
